requirements.txt
# Core Application Dependencies
streamlit>=1.28.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Note: asyncio is built into Python 3.7+, no need to install separately
